Shared code to create OpenVINO models. Changed README
